requirements.txt
keras==2.8.0
matplotlib==3.5.1
numpy==1.22.3
Pillow==9.0.1
requests==2.27.1
streamlit==1.8.0
tensorflow==2.8.0
